pyproject.toml
[project]
name = "ssago-bingo-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py>=2.6.4",
    "html2image>=2.0.7",
    "psycopg[binary]>=3.3.2",
]
